+++ src/pyclarify/__init__.py
@@ -18,4 +18,4 @@ from pyclarify.client import Client
 from pyclarify.views import Signal, SignalInfo, Item, ItemInfo, DataFrame
 import pyclarify.query
 
-__version__ = "0.4.0"
+__version__ = "0.4.0b2"