+++ src/pyclarify/__init__.py
@@ -18,5 +18,5 @@ from pyclarify.client import Client
 from pyclarify.views import Signal, SignalInfo, Item, ItemInfo, DataFrame
 import pyclarify.query
 
-__version__ = "0.4.0"
+__version__ = "0.4.1a1"
 __API_version__ = "1.1beta2"