+++ src/pyclarify/query/query.py
@@ -24,8 +24,6 @@ class DataQuery(BaseModel, extra=Extra.forbid):
     filter: Optional[Dict] = {}
     last: Optional[int] = -1
     rollup: Union[timedelta, Literal["window"]] = None
-    window_size: timedelta = None
-
 
 
 class ResourceQuery(BaseModel, extra=Extra.forbid):